+++ koishi/bot_utils/headers.py
@@ -1,0 +1,98 @@
+__all__ = ('get_header_for',)
+
+from random import random
+
+import config
+
+# https://patorjk.com/software/taag/#p=display&h=2&f=Doom&t=
+
+
+HEADER_KOISHI = (
+    '```\n'
+    ' _   __      _     _     _ \n'
+    '| | / /     (_)   | |   (_)\n'
+    '| |/ /  ___  _ ___| |__  _ \n'
+    '|    \ / _ \| / __| \'_ \| |\n'
+    '| |\  \ (_) | \__ \ | | | |\n'
+    '\_| \_/\___/|_|___/_| |_|_|\n'
+    '```'
+)
+
+
+HEADER_FLANDRE = (
+    '```\n'
+    '______ _                 _          \n'
+    '|  ___| |               | |         \n'
+    '| |_  | | __ _ _ __   __| |_ __ ___ \n'
+    '|  _| | |/ _` | \'_ \ / _` | \'__/ _ \\\n'
+    '| |   | | (_| | | | | (_| | | |  __/\n'
+    '\_|   |_|\__,_|_| |_|\__,_|_|  \___|\n'
+    '```'
+)
+
+HEADER_YOSHIKA = (
+    '```\n'
+    '__   __        _     _ _         \n'
+    '\ \ / /       | |   (_) |        \n'
+    ' \ V /___  ___| |__  _| | ____ _ \n'
+    '  \ // _ \/ __| \'_ \| | |/ / _` |\n'
+    '  | | (_) \__ \ | | | |   < (_| |\n'
+    '  \_/\___/|___/_| |_|_|_|\_\__,_|\n'
+    '```'
+)
+
+
+HEADER_KOISHI_EASTER_EGG = (
+    '```\n'
+    ' _____ __    ___ \n'
+    '|  ___/  |  /   |\n'
+    '|___ \`| | / /| |\n'
+    '    \ \| |/ /_| |\n'
+    '/\__/ /| |\___  |\n'
+    '\____/\___/   |_/\n'
+    '```'
+)
+
+HEADER_FLANDRE_EASTER_EGG = (
+    '```\n'
+    '______          _     _ _             \n'
+    '| ___ \        | |   | (_)            \n'
+    '| |_/ /   _  __| | __| |_ _ __   __ _ \n'
+    '|  __/ | | |/ _` |/ _` | | \'_ \ / _` |\n'
+    '| |  | |_| | (_| | (_| | | | | | (_| |\n'
+    '\_|   \__,_|\__,_|\__,_|_|_| |_|\__, |\n'
+    '                                 __/ |\n'
+    '                                |___/ \n'
+    '```'
+)
+
+
+HEADERS = {
+    config.KOISHI_ID: (HEADER_KOISHI, HEADER_KOISHI_EASTER_EGG),
+    config.FLANDRE_ID: (HEADER_FLANDRE, HEADER_FLANDRE_EASTER_EGG),
+    config.YOSHIKA_ID: (HEADER_YOSHIKA, HEADER_YOSHIKA),
+}
+
+HEADER_DEFAULT = HEADER_KOISHI
+
+
+def get_header_for(client):
+    """
+    Gets the header for the given client.
+    
+    Parameters
+    ----------
+    client : ``Client``
+        Client to get the header for.
+    
+    Returns
+    -------
+    header : `str`
+    """
+    header_choices = HEADERS.get(client.id, None)
+    if header_choices is None:
+        header = HEADER_DEFAULT
+    else:
+        header = header_choices[random() <= 0.01]
+    
+    return header