+++ koishi/plugins/embed_image_refresh/__init__.py
@@ -1,0 +1,6 @@
+from .refresh import *
+
+
+__all__ = (
+    *refresh.__all__,
+)