+++ koishi/plugins/sex/constants.py
@@ -8,6 +8,7 @@ SEX_IMAGES = [
     'https://cdn.discordapp.com/attachments/568837922288173058/1160496031331536936/chiruno-probably-sex.gif',
     'https://cdn.discordapp.com/attachments/568837922288173058/1160496033994915860/chiruno-yes-sex.gif',
     'https://cdn.discordapp.com/attachments/568837922288173058/1160496033399316540/chiruno-yes-sex-fast.gif',
+    'https://cdn.discordapp.com/attachments/568837922288173058/1183057597205590016/chiruno-totally-sex.gif',
     'https://cdn.discordapp.com/attachments/568837922288173058/1160496032468193400/chiruno-sex-2.png',
     'https://cdn.discordapp.com/attachments/568837922288173058/1160496030886936636/koishi-yes-sex.gif',
 ]